Level: object_to_follow and ugly mode
Instead of refer to the moto everywhere, the level object
now includes an object_to_follow function which will point
later on a ghost, a replay, a player.
Add the ugly mode. For the moment, only for the entities in the
master branch. The 'u' key switch from ugly, to uglyOver and normal
mode.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>