Ghost : replaying the ghost based on frame time, not frame count
Adding a frame time to frames.
It allows to save frames to different framerate.
It looks like a bug if you load the game, make a ghost a replay the level : during the 1st second, the ghost is doing nohting. In fact, the problem is that the start_time of the level is not when the page is finished to be loaded.
Level start time should be fixed to the end loading time.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>